Allows passing custom class names to MdInput (#411)
* Allows passing custom class names to MdInput

Enables the application of custom styles to the MdInput component by allowing a `className` prop.

This provides more flexibility for developers to customize the appearance of the input field without having to override the default styles.

* Passes otherProps to the input element.

Allows additional HTML attributes to be passed directly
to the underlying input element.
This provides greater flexibility when customizing
the input's behavior and appearance.

---------

Co-authored-by: Helge Johnsen <helge.johnsen@crayon.no>