Feature/53 search input (#416)
* feat(MdInputSearch): introduce new search input component

* feat(MdInputSearch): export MdInputSearch from package

* docs(MdInputSearch): add Storybook documentation

* Adds CSS for the Input Search component

This commit introduces the CSS file for the Input Search component,
defining the styles for its various states and elements, including:

- Layout and appearance
- States such as disabled, readonly, and error
- Prefix and suffix icons
- Help and support text

* feat: Add CSS import for InputSearch component

* refactor(MdInputSearch): Rename CSS classes to be component-specific

* feat(MdInputSearch): Implement search button using IconButton and IconSearch

* Displays search icon as prefix when no button

Ensures that the search icon is displayed as a prefix in the
input field when the button prop is not enabled. Also, adds
a class to handle the display of this prefix.

* Updates inputsearch HTML structure and styling

Updates the HTML structure for the inputsearch component, replacing `md-input` with `md-inputsearch` class prefixes to ensure correct styling and component separation.

Removes the suffix container in favor of using an IconButton
with a search icon to trigger the search action.

* Removes deprecated mode warning

Removes the console warning for the deprecated "normal" mode of the MdInputSearch component.

* Renames class for inputsearch

Updates the class name in the inputsearch example to align with the component's naming convention.

* Removes deprecated input mode

Removes the deprecated "normal" mode from the input component, encouraging the use of "medium" instead.

Also, explicitly sets the input type to "search" for improved semantics.

* Removes unused InputSearch properties

Removes the `type` and `readOnly` properties from the InputSearch component's Storybook configuration.

These properties are deemed unnecessary and are removed to simplify the component's interface and reduce potential confusion for users.

* Updates InputSearch story to use correct README

Updates the InputSearch story to import the README file from the correct directory, ensuring that the component documentation is accurately displayed.

---------

Co-authored-by: Helge Johnsen <helge.johnsen@crayon.no>