Adds labels prop for accessibility and i18n
Introduces a `labels` prop to `MdInputSearch` and `MdInfoTag`, and `MdAlertMessage` components.

This change allows customization of aria-labels for improved accessibility and internationalization, providing default labels with override capability.

The `labels` prop accepts an object to override the default aria-labels used in the components. This enhancement allows developers to easily adapt the components to different languages and user needs, improving the overall user experience.