Allow all arialabels to be replaced by a string property to allow for external translations (#425)
* Adds customizable labels to FileList component

Allows customization of aria-labels for action buttons in the FileList component.

This provides improved accessibility and internationalization by enabling users to define their own labels for the delete, download, edit, and print actions. Falls back to Norwegian defaults if custom labels aren't provided.

* Adds default and customizable labels to file list

Ensures default labels for file actions (download, delete, edit, print) are provided.

Allows customization of these labels via the `labels` prop,
merging provided labels with the defaults.
This improves accessibility and internationalization support.

* Adds customizable labels to form elements

Adds a `labels` prop to various form elements (Input, TextArea, CheckboxGroup, RadioGroup, MultiSelect, MultiAutocomplete, ComboBox, ComboBoxGrouped) and HelpButton to allow customization of aria-labels.
A default set of Norwegian labels is provided if custom labels are not specified.
This enhances accessibility by allowing developers to provide more context-specific labels for users.

* Adds labels prop for accessibility and i18n

Introduces a `labels` prop to `MdInputSearch` and `MdInfoTag`, and `MdAlertMessage` components.

This change allows customization of aria-labels for improved accessibility and internationalization, providing default labels with override capability.

The `labels` prop accepts an object to override the default aria-labels used in the components. This enhancement allows developers to easily adapt the components to different languages and user needs, improving the overall user experience.

* Replaces aria-label with labels prop in MdIconButton

Deprecates the `aria-label` prop in `MdIconButton` and introduces a `labels` prop to provide separate labels for the button and tooltip.

This change provides more flexibility and clarity in defining labels for accessibility, addressing potential ambiguity when the same label is used for both the button and its tooltip.
A warning message is displayed if the deprecated `aria-label` is used to encourage migration to the new `labels` prop.

* Replaces `aria-label` with `label` prop in MdIconButton

Deprecates the `aria-label` prop in favor of a single `label`
prop for both button aria-label and tooltip content.

This simplifies the component API and reduces potential
confusion regarding label management.

Adds a console warning to inform users about the deprecation.

* Removes deprecated aria-label prop from IconButton

Removes the deprecated `aria-label` prop from the `MdIconButton` component and updates the component to ensure `aria-label` is not passed directly via other props.

This change enforces the use of the `label` prop for accessibility purposes and provides a console warning if `aria-label` is still being used.

---------

Co-authored-by: Helge Johnsen <helge.johnsen@crayon.no>