Adds onSearch functionality to InputSearch (#422)
* Adds onSearch functionality to InputSearch

Enables the InputSearch component to trigger a search action when the user submits a search term, either by pressing Enter or clicking the search button.
Also keeps the search term in state.

* Renames Storybook title for clarity

Updates the Storybook title to accurately reflect the component being showcased, improving discoverability and organization within the Storybook interface.

* Improves search input handling and initialization

Enhances the search input component by:

- Adding `onSearch` as optional.
- Submitting the search term when the Enter key is pressed.
- Initializing the search term from the `value` prop for controlled components.

* Added description

* Suggest implementation for MdSearchInput

* Refactor MdInputSearch to handle form submission and search term retrieval

* Improve form submission and search value handling

* Refactors InputSearch story for better clarity

Updates the InputSearch story to utilize the `args` property
directly on the story export.

This approach improves readability and simplifies the story
definition, making it easier to understand and maintain.

---------

Co-authored-by: Helge Johnsen <helge.johnsen@crayon.no>
Co-authored-by: Aurora Scharff <aurora.sofie@gmail.com>