New Tag, deprecate InfoTag. (#439)
* Initial code for new tag

* Adds tag component enhancements

Improves the tag component by adding icon support and tooltip functionality.

The changes include:
- Adding a `customIcon` prop to display icons within the tag.
- Implementing a `tooltipOnly` prop, when activated a tooltip will appear when hovering over the custom icon.

The styling is updated to include spacing for the icon.

* Adds icon support to the MdTag component

Introduces the ability to display icons within the MdTag component based on the theme and a new `showIcon` property.

This change allows for visual differentiation of tags using predefined icons for "success", "warning", "info", and "error" themes, improving user experience. It also allows for overriding the predefined icons with the `customIcon` property.

Also refactors how `customIcon` property worked together with `tooltipOnly` property.

* Deprecates MdInfoTag component

Marks the MdInfoTag component as deprecated by:

- Adding a console warning to inform developers about the deprecation and suggest using MdTag instead.
- Moving the component's story to a "Deprecated" section in Storybook to visually indicate its deprecated status.
- Updating the Storybook documentation to highlight the component's deprecated status.

* Removes unnecessary console warning

Removes a console warning that was being displayed
when `tooltipOnly` was enabled without a `customIcon`.
This check is no longer needed.

---------

Co-authored-by: Helge Johnsen <helge.johnsen@crayon.no>