Misc bugfix, mostly misspelling of variables (#440)
* Uses label prop for button text

Replaces the `aria-label` prop with the `label` prop for buttons within the file list component.

This change ensures that the button text is properly rendered and accessible, improving the user experience.

* Fixes typo in MdInputSearchProps interface

Corrects the spelling of the 'labels' property in the MdInputSearchProps interface, resolving a typo.

* Replaces `aria-label` with `label` prop

Uses the `label` prop instead of `aria-label` for the `MdIconButton` component in `MdInputSearch` and `MdAlertMessage`.

This change aligns the component's API for accessibility with the design system's guidelines.

---------

Co-authored-by: Helge Johnsen <helge.johnsen@crayon.no>